Cloud/data_subscriber/requirements.txt
setuptools
paho-mqtt
pika